+++ magick/pixel-private.h
@@ -41,8 +41,8 @@ static inline MagickBooleanType IsGrayPixel(const PixelPacket *pixel)
       alpha,
       beta;
 
-    alpha=GetPixelRed(pixel)-(double) GetPixelGreen(pixel);
-    beta=GetPixelGreen(pixel)-(double) GetPixelBlue(pixel);
+    alpha=(double) GetPixelRed(pixel)-(double) GetPixelGreen(pixel);
+    beta=(double) GetPixelGreen(pixel)-(double) GetPixelBlue(pixel);
     if ((fabs(alpha) <= MagickEpsilon) && (fabs(beta) <= MagickEpsilon))
       return(MagickTrue);
   }
@@ -64,10 +64,10 @@ static inline MagickBooleanType IsMonochromePixel(const PixelPacket *pixel)
       alpha,
       beta;
 
-    alpha=GetPixelRed(pixel)-(double) GetPixelGreen(pixel);
-    beta=GetPixelGreen(pixel)-(double) GetPixelBlue(pixel);
+    alpha=(double) GetPixelRed(pixel)-(double) GetPixelGreen(pixel);
+    beta=(double) GetPixelGreen(pixel)-(double) GetPixelBlue(pixel);
     if (((fabs((double) GetPixelRed(pixel)) <= MagickEpsilon) ||
-         (fabs((double) GetPixelRed(pixel)-QuantumRange) <= MagickEpsilon)) &&
+         (fabs((double) GetPixelRed(pixel)-(double) QuantumRange) <= MagickEpsilon)) &&
         (fabs(alpha) <= MagickEpsilon) && (fabs(beta) <= MagickEpsilon))
       return(MagickTrue);
     }
